relocate cola-analysis
